refactor(api): remove same-release deprecations + Reviewer A nits
v0.20 was shipping ~7 APIs that were introduced in this very release
and then immediately marked #[deprecated] — pure API tax with no
migration benefit. Drop entirely. Plus several nits flagged by
Reviewer A that don't deserve their own commit.

Removed:
- Context::gauge_w, gauge_colored, line_gauge_with, breadcrumb_sep
- LineGaugeOpts struct
- HighlightRange::single (kept ::line)
- Gauge::label_owned, LineGauge::label_owned
- _prev_focus_rects dead field on Context

Refactored:
- Gauge::label / LineGauge::label now take impl Into<String> so
  callers with owned Strings don't pay a redundant clone path
- slt_warn macro: ANSI escape sequences use \x1b notation in source

Demo cleanups:
- v020_split_pane.rs: drop redundant f64::from(r.ratio) cast
- v020_regression_panel.rs: use theme.surface / theme.primary /
  theme.text_dim instead of hardcoded Color::Rgb(40, 40, 40), so
  the demo also showcases the theme system (#226)

Test cleanups:
- Removed deprecated-shim assertions in tests/v020_widgets.rs
  (gauge_w, line_gauge_with, HighlightRange::single alias) and
  tests/widgets.rs (breadcrumb_sep)
- Migrated remaining HighlightRange::single sites to ::line in
  tests, examples, and gutter doctest

CHANGELOG.md cleanup:
- Removed deprecation-shim sentences from API consistency entry
- Migration recipe phrased for v0.19 preview users only
- LineGaugeOpts public-during-deprecation sentence removed

+++ examples/v020_split_pane.rs
@@ -106,8 +106,7 @@ pub fn render(ui: &mut Context, state: &mut DemoState) {
 
             ui.text(format!(
                 "ratio = {:.2}    drag_active = {}",
-                f64::from(r.ratio),
-                r.drag_active
+                r.ratio, r.drag_active
             ))
             .dim();
             ui.text("Ctrl-Q / Esc quits.").dim();
